requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
Flask-Talisman==1.1.0
Flask-Limiter==3.7.0
Flask-Caching==2.3.0
Flask-Compress==1.14
limits==3.13.0
argon2-cffi==23.1.0
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
APScheduler==3.10.4
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.4.0
zxcvbn-python==4.4.24
pytest==8.3.3
pytest-flask==1.3.0
openpyxl==3.1.5
gunicorn==21.2.0
Werkzeug==3.0.3
reportlab==4.2.0
feedparser==6.0.10
requests==2.31.0
cryptography==43.0.3
bleach==6.1.0
python-magic==0.4.27
psutil==5.9.8
pandas==2.2.0
matplotlib==3.8.4
diff-match-patch==20230430
playwright==1.55.0
